Modifications to grammar to be more compatible with KaSim Addition of mean and standard deviation chart for compartments Updated record and replay file formats to match plot data format of KaSim

+++ src/java/org/demonsoft/spatialkappa/model/Compartment.java
@@ -99,7 +99,7 @@ public class Compartment {
     }
 
     public Serializable createValueArray() {
-        Serializable result = (Serializable) Array.newInstance(int.class, dimensions);
+        Serializable result = (Serializable) Array.newInstance(float.class, dimensions);
         return result;
     }
